Added fix fir _FillValue from abdul.g.khan@nasa.gov
This is a fix for issue ODIP-133 where an unsigned byte
array is promoted to a short and the type of the _FillValue
attribtue is not. This results in a type-mismatch for the
attribute and that makes netcdf clients gag. Thanks to Abdul
for the fix. We should check other types for this same issue,
although unsigned byte maybe a special case...