New tests usinf in place of the older ones.
The new tests in fileout_netcdf use the handler while the older
tests used code that that simulated the handler's operations. It
was OK before teh handler was modified several times... The new
tests use the dapreader handler except for two tests that use
hdf4 and hdf5 - the latter can be factored out with a bit more work.