Tweaked the compression flag; ByteToShort is not used
The ByteToShort key might not be very useful. I don't see it being
used at all with DAP2 - the byte type in netcdf is signed and in
DAP2 is unsigned. The code added by Abdul actually transforms the
type of an associated attribute so that netCDF client software won't
fail when it detects a type mismatch between the variable and its
_FillValue attribute. For DAP4 this key might be used, so I left it
in the code.