Updated the baselines so that they filter out date-time information
Nathan added support for the CF history attribute (!) but that
includes a date-time string (according to the spec) and that breaks
tests (since time is passing...). We now filter out that date-time
string from the baselines and replace it with the phrase
'removed date-time'. This updates the test code and the baselies.